Refactor RabbitMQ types package
