+++ pkg/cmd/server/main.go
@@ -28,9 +28,9 @@ import (
 	"github.com/scorify/scorify/pkg/graph"
 	"github.com/scorify/scorify/pkg/graph/directives"
 	"github.com/scorify/scorify/pkg/rabbitmq/management"
-	"github.com/scorify/scorify/pkg/rabbitmq/management/types"
 	"github.com/scorify/scorify/pkg/rabbitmq/management/vhosts"
 	"github.com/scorify/scorify/pkg/rabbitmq/rabbitmq"
+	"github.com/scorify/scorify/pkg/rabbitmq/types"
 	"github.com/scorify/scorify/pkg/structs"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
