+++ pkg/rabbitmq/management/users/get.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/scorify/scorify/pkg/rabbitmq/management/types"
+	"github.com/scorify/scorify/pkg/rabbitmq/types"
 )
 
 func (c *UsersClient) Get(name string) (*userResponse, *types.ErrorResponse, error) {
