+++ pkg/rabbitmq/management/vhosts/delete.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/scorify/scorify/pkg/rabbitmq/management/types"
+	"github.com/scorify/scorify/pkg/rabbitmq/types"
 )
 
 func (c *VhostsClient) Delete(vhost string) (*types.ErrorResponse, error) {
