+++ pkg/rabbitmq/management/vhosts/get.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"net/url"
 
-	"github.com/scorify/scorify/pkg/rabbitmq/management/types"
+	"github.com/scorify/scorify/pkg/rabbitmq/types"
 )
 
 func (c *VhostsClient) Get(vhost string) (*vhostsResponse, *types.ErrorResponse, error) {
