+++ pkg/rabbitmq/management/vhosts/list.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/scorify/scorify/pkg/rabbitmq/management/types"
+	"github.com/scorify/scorify/pkg/rabbitmq/types"
 )
 
 func (c *VhostsClient) List() ([]*vhostsResponse, *types.ErrorResponse, error) {
