+++ pkg/rabbitmq/management/vhosts/put.go
@@ -8,7 +8,7 @@ import (
 	"net/url"
 	"strings"
 
-	"github.com/scorify/scorify/pkg/rabbitmq/management/types"
+	"github.com/scorify/scorify/pkg/rabbitmq/types"
 )
 
 func (c *VhostsClient) Put(name string, description string, tags []string, defaultQueueType QueueType) (*types.ErrorResponse, error) {
