+++ pkg/rabbitmq/types/main.go
